# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Duolingo Clone API",
    "description": "Duolingo clone backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Me",
        "description": "Retrieve current authenticated user state, hearts, streak, and daily progress.",
        "operationId": "get_me_api_me_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Profile",
        "description": "Retrieve full user profile with stats and recent activity.",
        "operationId": "get_profile_api_profile_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Settings",
        "description": "Retrieve user preferences.",
        "operationId": "get_settings_api_settings_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update Settings",
        "description": "Update user preferences.",
        "operationId": "update_settings_api_settings_patch",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lessons/{lesson_id}": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "summary": "Get Lesson",
        "description": "Retrieve lesson metadata and public exercises. answer_json is NEVER included.",
        "operationId": "get_lesson_api_lessons__lesson_id__get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lesson_id}/start": {
      "post": {
        "tags": [
          "Lessons"
        ],
        "summary": "Start Lesson",
        "description": "Authoritatively initiate a lesson attempt.\n\nRules:\n- 409 OUT_OF_HEARTS if user has 0 hearts.\n- 403 LESSON_LOCKED if lesson is locked.\n- Abandons existing in_progress attempt if one exists.\n- Exercises delivered in deterministic exercise_order without answers.",
        "operationId": "start_lesson_api_lessons__lesson_id__start_post",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hearts/refill": {
      "post": {
        "tags": [
          "Gamification"
        ],
        "summary": "Refill Hearts",
        "description": "Refill hearts back to 5 via gems (350) or practice (free).",
        "operationId": "refill_hearts_api_hearts_refill_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefillHeartsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefillHeartsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/leaderboard": {
      "get": {
        "tags": [
          "Gamification"
        ],
        "summary": "Get Leaderboard",
        "description": "Get current week's leaderboard standings.",
        "operationId": "get_leaderboard_api_leaderboard_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/achievements": {
      "get": {
        "tags": [
          "Gamification"
        ],
        "summary": "List Achievements",
        "description": "List achievements and user unlocking milestones.",
        "operationId": "list_achievements_api_achievements_get",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementItem"
                  },
                  "title": "Response List Achievements Api Achievements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AchievementItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "badge_icon": {
            "type": "string",
            "title": "Badge Icon"
          },
          "target_value": {
            "type": "integer",
            "title": "Target Value"
          },
          "current_value": {
            "type": "integer",
            "title": "Current Value"
          },
          "is_unlocked": {
            "type": "boolean",
            "title": "Is Unlocked"
          },
          "unlocked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unlocked At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "description",
          "badge_icon",
          "target_value",
          "current_value",
          "is_unlocked"
        ],
        "title": "AchievementItem"
      },
      "ActivityItem": {
        "properties": {
          "activity_date": {
            "type": "string",
            "title": "Activity Date"
          },
          "xp_earned": {
            "type": "integer",
            "title": "Xp Earned"
          },
          "lessons_completed": {
            "type": "integer",
            "title": "Lessons Completed"
          }
        },
        "type": "object",
        "required": [
          "activity_date",
          "xp_earned",
          "lessons_completed"
        ],
        "title": "ActivityItem"
      },
...
      "ExerciseClientView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exercise_order": {
            "type": "integer",
            "title": "Exercise Order"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "exercise_order",
          "type",
          "payload"
        ],
        "title": "ExerciseClientView"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeaderboardRanking": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "xp_earned": {
            "type": "integer",
            "title": "Xp Earned"
          },
          "is_current_user": {
            "type": "boolean",
            "title": "Is Current User"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rank",
          "username",
          "xp_earned",
          "is_current_user"
        ],
        "title": "LeaderboardRanking"
      },
      "LeaderboardResponse": {
        "properties": {
          "week_start_date": {
            "type": "string",
            "title": "Week Start Date"
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardRanking"
            },
            "type": "array",
            "title": "Rankings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "week_start_date",
          "rankings"
        ],
        "title": "LeaderboardResponse"
      },
      "LessonMetadataResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseClientView"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "skill_id",
          "title",
          "exercises"
        ],
        "title": "LessonMetadataResponse"
      },
...
      "ProfileResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserProfileStats"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            },
            "type": "array",
            "title": "Recent Activity"
          },
          "unlocked_achievements_count": {
            "type": "integer",
            "title": "Unlocked Achievements Count"
          }
        },
        "type": "object",
        "required": [
          "user",
          "recent_activity",
          "unlocked_achievements_count"
        ],
        "title": "ProfileResponse"
      },
      "RefillHeartsRequest": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "gems",
              "practice"
            ],
            "title": "Method"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "method"
        ],
        "title": "RefillHeartsRequest"
      },
      "RefillHeartsResponse": {
        "properties": {
          "hearts": {
            "type": "integer",
            "title": "Hearts"
          },
          "gems": {
            "type": "integer",
            "title": "Gems"
          },
          "hearts_updated_at": {
            "type": "string",
            "title": "Hearts Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hearts",
          "gems",
          "hearts_updated_at"
        ],
        "title": "RefillHeartsResponse"
      },
      "SettingsResponse": {
        "properties": {
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "date_offset_days": {
            "type": "integer",
            "title": "Date Offset Days"
          }
        },
        "type": "object",
        "required": [
          "timezone",
          "date_offset_days"
        ],
        "title": "SettingsResponse"
      },
      "SettingsUpdateRequest": {
        "properties": {
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "SettingsUpdateRequest"
      },
...
      "StartAttemptResponse": {
        "properties": {
          "attempt_id": {
            "type": "integer",
            "title": "Attempt Id"
          },
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "hearts_remaining": {
            "type": "integer",
            "title": "Hearts Remaining"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseClientView"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "lesson_id",
          "status",
          "hearts_remaining",
          "exercises"
        ],
        "title": "StartAttemptResponse"
      },
...
      "UserProfileStats": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "xp_total": {
            "type": "integer",
            "title": "Xp Total"
          },
          "streak_current": {
            "type": "integer",
            "title": "Streak Current"
          },
          "total_crowns": {
            "type": "integer",
            "title": "Total Crowns"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "created_at",
          "xp_total",
          "streak_current",
          "total_crowns"
        ],
        "title": "UserProfileStats"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "hearts": {
            "type": "integer",
            "title": "Hearts"
          },
          "max_hearts": {
            "type": "integer",
            "title": "Max Hearts",
            "default": 5
          },
          "hearts_updated_at": {
            "type": "string",
            "title": "Hearts Updated At"
          },
          "next_heart_in_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Heart In Seconds"
          },
          "next_heart_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Heart At"
          },
          "xp_total": {
            "type": "integer",
            "title": "Xp Total"
          },
          "gems": {
            "type": "integer",
            "title": "Gems"
          },
          "streak_current": {
            "type": "integer",
            "title": "Streak Current"
          },
          "streak_active_today": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Streak Active Today",
            "default": false
          },
          "daily_xp_goal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Xp Goal",
            "default": 20
          },
          "daily_xp_progress": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Xp Progress",
            "default": 0
          },
          "date_offset_days": {
            "type": "integer",
            "title": "Date Offset Days"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "hearts",
          "hearts_updated_at",
          "xp_total",
          "gems",
          "streak_current",
          "date_offset_days",
          "timezone"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}